# Use MaybeUninit in VecDeque to remove the undefined behavior of slice