@JmPotato 2022-03-09
> Nice, thank you for working on this! I left one small comment, but otherwise this looks good to me.
> 
> > Since I haven't received any feedback for a while, [..]
> 
> Unfortunately we're all quite busy and there's a lot of PRs to review, so it's not uncommon for a PR to sit still for a week or two. Unless it's a very critical issue, two weeks is probably a good time to send a ping or reassign it.

Thanks for the explanation! It seems that I am too eager, xD. I have addressed the comment. PTAL.