@JmPotato 2022-03-10
> @bors r- failed in a rollup it seems: [#94784 (comment)](https://github.com/rust-lang/rust/pull/94784#issuecomment-1063424379) @bors rollup=never (perf change)

@matthiaskrgr @m-ou-se Hi, I updated the `Debug` implementation of `VecDeque` to keep compatible with the previous behavior. This should fix the falling test, PTAL again, thx!