@JmPotato 2022-03-11
A memory leak is found during my own test, it turns out that dropping a `MaybeUninit<T>` will never call `T`’s drop code. It is my responsibility to make sure `T` gets dropped if it got initialized. I will try to fix it later. ref [here](https://doc.rust-lang.org/stable/std/mem/union.MaybeUninit.html#impl).

```shell
diff of stdout:

 [2, 2] Iter([2, 2], [])
-Iter([], [])
+Iter([2, 2], [])


The actual stdout differed from the expected stdout.
Actual stdout saved to /var/folders/x_/z7jrp4m90z948c1n3qnzfmt80000gn/T/compiletestHn7VSL/vecdeque.stage-id.stdout
normalized stderr:
The following memory was leaked: alloc4102 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc4151 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc8950 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc8999 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}

error: the evaluated program leaked memory

note: pass `-Zmiri-ignore-leaks` to disable this check

error: aborting due to previous error




The actual stderr differed from the expected stderr.
Actual stderr saved to /var/folders/x_/z7jrp4m90z948c1n3qnzfmt80000gn/T/compiletestHn7VSL/vecdeque.stage-id.stderr
To update references, rerun the tests and pass the `--bless` flag
To only update this specific test, also pass `--test-args vecdeque.rs`

error: 2 errors occurred comparing output.
status: exit status: 1
command: "/Users/jmpotato/Code/rust/build/aarch64-apple-darwin/stage1-tools-bin/miri" "tests/run-pass/vecdeque.rs" "-L" "/var/folders/x_/z7jrp4m90z948c1n3qnzfmt80000gn/T/compiletestHn7VSL" "--target=aarch64-apple-darwin" "--error-format" "json" "-C" "prefer-dynamic" "-o" "/var/folders/x_/z7jrp4m90z948c1n3qnzfmt80000gn/T/compiletestHn7VSL/vecdeque.stage-id" "-A" "unused" "--edition" "2018" "-Astable-features" "--sysroot" "/Users/jmpotato/Library/Caches/org.rust-lang.miri/HOST" "-Zmiri-tag-raw-pointers" "-L" "/var/folders/x_/z7jrp4m90z948c1n3qnzfmt80000gn/T/compiletestHn7VSL/vecdeque.stage-id.aux"
stdout:
------------------------------------------
[2, 2] Iter([2, 2], [])
Iter([2, 2], [])

------------------------------------------
stderr:
------------------------------------------
The following memory was leaked: alloc4102 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc4151 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc8950 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}
alloc8999 (Rust heap, size: 4, align: 4) {
    02 00 00 00                                     │ ....
}

{"message":"the evaluated program leaked memory","code":null,"level":"error","spans":[],"children":[],"rendered":"error: the evaluated program leaked memory\n\n"}
{"message":"pass `-Zmiri-ignore-leaks` to disable this check","code":null,"level":"note","spans":[],"children":[],"rendered":"note: pass `-Zmiri-ignore-leaks` to disable this check\n\n"}
{"message":"aborting due to previous error","code":null,"level":"error","spans":[],"children":[],"rendered":"error: aborting due to previous error\n\n"}

------------------------------------------

test [ui] run-pass/vecdeque.rs ... FAILED
```